added product bg images
