+++ src/app/[locale]/products/[product]/page.tsx
@@ -1,11 +1,10 @@
 import React from "react";
-import { useTranslations } from "next-intl";
-import Breadcrumbs from "@/components/shared/breadcrumbs/Breadcrumbs";
 import { Locale } from "@/types/locale";
 import { productsList } from "@/mockedData/products";
 import NotFound from "../../not-found";
 import SimilarProducts from "@/components/product/similarProducts/SimilarProducts";
 import { ProductItem } from "@/types/productItem";
+import ProductInfo from "@/components/product/productInfo/ProductInfo";
 
 export default function ProductPage({
   params,
@@ -15,8 +14,6 @@ export default function ProductPage({
     locale: Locale;
   };
 }) {
-  const t = useTranslations();
-
   const { product, locale } = params;
 
   const currentProduct = productsList[locale]?.find(
@@ -28,23 +25,9 @@ export default function ProductPage({
     return <NotFound />;
   }
 
-  const { title, volume, trademark, category } = currentProduct;
-
-  const crumbs = [
-    { label: t("breadcrumbs.home"), href: "/" },
-    { label: t("breadcrumbs.products"), href: "/products" },
-    {
-      label:
-        `${title} ${volume} ${t("productsPage.ml")} ${t(
-          "productsPage.trademark"
-        )} ${trademark.title}` || "",
-      href: `/products/${trademark.id}-${category}-${volume}` || "",
-    },
-  ];
-
   return (
     <>
-      <Breadcrumbs crumbs={crumbs} />
+      <ProductInfo currentProduct={currentProduct} />
       <SimilarProducts currentProduct={currentProduct} />
     </>
   );
