+++ src/components/product/productInfo/ProductBgImages.tsx
@@ -1,0 +1,30 @@
+import Image from "next/image";
+import React from "react";
+
+export default function ProductBgImages() {
+  return (
+    <>
+      <Image
+        src="/images/backgrounds/productInfo/productBgMob.svg"
+        alt="grey background"
+        width="360"
+        height="316"
+        className="tab:hidden absolute top-0 left-0 -z-10 w-full h-auto"
+      />
+      <Image
+        src="/images/backgrounds/productInfo/productBgTab.svg"
+        alt="grey background"
+        width="768"
+        height="362"
+        className="hidden tab:block laptop:hidden absolute top-0 left-0 -z-10 w-full h-auto"
+      />
+      <Image
+        src="/images/backgrounds/productInfo/productBgDesk.svg"
+        alt="grey background"
+        width="1280"
+        height="366"
+        className="hidden laptop:block absolute top-0 left-0 -z-10 w-full h-auto"
+      />
+    </>
+  );
+}
